Truly syncronous game creation.
Meaning, I spent a lot of time to make the list modification theoretically resistant
to race condition. Half of a kingdom and a horse for usual, stupid, plain lock.
But the best thing? For an hour I tried to understand why doesn't my for
evluates. Or map, for that matter. Finally, I understood. Because it's lazy!
I need some sleep.